Help WP find featured image from post
Introduces possible bug/feature:
If image is uploaded into post, and then deleted (but still attached to post), it might be the first attachment and therefore selected as featured image even though it is not in the post any more. This is a speed tradeoff. Simple solution to bug: select featured image as you should.. :)